Protect routes using token passed in headers and update tests to reflect the change

+++ backend/tests/test_patient_integration.py
@@ -1,25 +1,25 @@
-def test_patients_get_empty(client):
-  response = client.get('/patients/')
+def test_patients_get_empty(client,access_token):
+  response = client.get('/patients/',headers={'Authorization': 'Bearer '+access_token})
   assert response.status_code == 200
   assert response.json == []
 
-def test_patients_get_not_empty(client,populate_database):
-  response = client.get('/patients/')
+def test_patients_get_not_empty(client,populate_database,access_token):
+  response = client.get('/patients/',headers={'Authorization': 'Bearer '+access_token})
   assert response.status_code == 200
   assert response.json != []
 
-def test_patient_get_empty(client):
-  response = client.get('/patients/1')
+def test_patient_get_empty(client,access_token):
+  response = client.get('/patients/1',headers={'Authorization': 'Bearer '+access_token})
   assert response.status_code == 422
   assert response.json == {'error': 'Patient does not exist'}
 
-def test_patient_get_not_empty(client,populate_database):
-  response = client.get('/patients/1')
+def test_patient_get_not_empty(client,populate_database,access_token):
+  response = client.get('/patients/1',headers={'Authorization': 'Bearer '+access_token})
   assert response.status_code == 200
   assert response.json['first_name'] == 'Test'
   assert response.json['last_name'] == 'Patient'
 
-def test_patient_create_valid(client,populate_database):
+def test_patient_create_valid(client,populate_database,access_token):
   data = {
     'first_name': 'Test',
     'last_name': 'Patient2',
@@ -27,12 +27,12 @@ def test_patient_create_valid(client,populate_database):
     'physician_first_name': 'Test',
     'physician_last_name': 'Physician'
   }
-  response = client.post('/patients/', json=data)
+  response = client.post('/patients/', json=data,headers={'Authorization': 'Bearer '+access_token})
   assert response.status_code == 200
   assert response.json['first_name'] == 'Test'
   assert response.json['last_name'] == 'Patient2'
 
-def test_patient_create_duplicate(client,populate_database):
+def test_patient_create_duplicate(client,populate_database,access_token):
   data = {
     'first_name': 'Test',
     'last_name': 'Patient',
@@ -40,11 +40,11 @@ def test_patient_create_duplicate(client,populate_database):
     'physician_first_name': 'Test',
     'physician_last_name': 'Physician'
   }
-  response = client.post('/patients/', json=data)
+  response = client.post('/patients/', json=data,headers={'Authorization': 'Bearer '+access_token})
   assert response.status_code == 422
   assert response.json == {'error': 'Patient already exists'}
 
-def test_patient_create_invalid_physician(client,populate_database):
+def test_patient_create_invalid_physician(client,populate_database,access_token):
   data = {
     'first_name': 'Test',
     'last_name': 'Patient2',
@@ -52,38 +52,38 @@ def test_patient_create_invalid_physician(client,populate_database):
     'physician_first_name': 'Invalid',
     'physician_last_name': 'Physician'
   }
-  response = client.post('/patients/', json=data)
+  response = client.post('/patients/', json=data,headers={'Authorization': 'Bearer '+access_token})
   assert response.status_code == 422
   assert response.json == {'error': 'Physician does not exist'}
 
-def test_patient_update_valid(client,populate_database):
+def test_patient_update_valid(client,populate_database,access_token):
   data = {
     'first_name': 'Test',
     'last_name': 'Patient',
     'email_address': 'new@test.com',
     'physician_id': 1
   }
-  response = client.put('/patients/1', json=data)
+  response = client.put('/patients/1', json=data,headers={'Authorization': 'Bearer '+access_token})
   assert response.status_code == 200
   assert response.json['email_address'] == 'new@test.com'
 
-def test_patient_update_invalid(client):
+def test_patient_update_invalid(client,access_token):
   data = {
     'first_name': 'Test',
     'last_name': 'Patient',
     'email_address': 'new@test.com',
     'physician_id': 1
   }
-  response = client.put('/patients/1', json=data)
+  response = client.put('/patients/1', json=data,headers={'Authorization': 'Bearer '+access_token})
   assert response.status_code == 422
   assert response.json == {'error': 'Patient does not exist'}
 
-def test_patient_delete_valid(client,populate_database):
-  response = client.delete('/patients/1')
+def test_patient_delete_valid(client,populate_database,access_token):
+  response = client.delete('/patients/1',headers={'Authorization': 'Bearer '+access_token})
   assert response.status_code == 200
   assert response.json == {'message': 'Patient deleted successfully'}
 
-def test_patient_delete_invalid(client):
-  response = client.delete('/patients/1')
+def test_patient_delete_invalid(client,access_token):
+  response = client.delete('/patients/1',headers={'Authorization': 'Bearer '+access_token})
   assert response.status_code == 422
   assert response.json == {'error': 'Patient does not exist'}
